requirements.txt
streamlit==1.31.1
pandas==2.1.3
google-cloud-bigquery==3.13.0
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
